Moved io from utils to a submodule
