+++ mlcolvar/io/_utils.py
