+++ mlcolvar/io/graphs/_utils.py
